Move BlockProvider to separate file
